requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
bcrypt==4.0.1
mysql-connector-python==8.3.0
